fix(template): fixed typo in Cargo.toml
There was a typo in the Cargo.toml file that would create a completely broken project. First things users would notice when doing steel new is that the rust analyzer will complain, and the compilation will fail.

This PR fixes that.

I guess a new version of the CLI would be required.